# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tool Rent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List shops
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopListResponse"
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a shop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopRequest"
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get shop by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Shop info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    patch:
      operationId: updateShop
      tags: [shops]
      summary: Update shop
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopRequest"
      responses:
        "200":
          description: Updated shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /rentals:
    get:
      operationId: listRentals
      tags: [rentals]
      summary: List rentals
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, returned, overdue]
        - name: customerId
          in: query
          schema:
            type: number
        - name: shopId
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of rentals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalListResponse"
    post:
      operationId: createRental
      tags: [rentals]
      summary: Start a rental
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRentalRequest"
      responses:
        "201":
          description: Rental created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /rentals/{id}:
    get:
      operationId: getRental
      tags: [rentals]
      summary: Get rental by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Rental info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /rentals/{id}/return:
    post:
      operationId: returnRental
      tags: [rentals]
      summary: Return a rented tool
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnRentalRequest"
      responses:
        "200":
          description: Rental returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /rentals/start-by-qr:
    post:
      operationId: startRentalByQr
      tags: [rentals]
      summary: Start rental by scanning QR code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRentalByQrRequest"
      responses:
        "201":
          description: Rental started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /rentals/return-by-qr:
    post:
      operationId: returnRentalByQr
      tags: [rentals]
      summary: Return rental by scanning QR code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnByQrRequest"
      responses:
        "200":
          description: Rental returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Shop:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        ownerId:
          type: number
        subscriptionStatus:
          type: string
          enum: [active, expired, trial]
        subscriptionEndsAt:
          type: string
          format: date-time
        commission:
          type: number
          description: Platform commission percentage
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - address
        - phone
        - ownerId
        - subscriptionStatus
        - commission
        - isActive
        - createdAt
    CreateShopRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        ownerId:
          type: number
        commission:
          type: number
      required:
        - name
        - address
        - phone
        - ownerId
    UpdateShopRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        commission:
          type: number
        isActive:
          type: boolean
        subscriptionStatus:
          type: string
          enum: [active, expired, trial]
    ShopListResponse:
      type: object
      properties:
        shops:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - shops
        - total
        - page
        - limit
...
    Rental:
      type: object
      properties:
        id:
          type: number
        toolId:
          type: number
        toolName:
          type: string
        customerId:
          type: number
        customerName:
          type: string
        shopId:
          type: number
        workerId:
          type: number
        status:
          type: string
          enum: [active, returned, overdue]
        rentalPrice:
          type: number
        depositAmount:
          type: number
        totalAmount:
          type: number
        startedAt:
          type: string
          format: date-time
        returnedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        damageNote:
          type: string
        damageCost:
          type: number
      required:
        - id
        - toolId
        - customerId
        - shopId
        - status
        - rentalPrice
        - depositAmount
        - totalAmount
        - startedAt
    CreateRentalRequest:
      type: object
      properties:
        toolId:
          type: number
        customerId:
          type: number
        dueDate:
          type: string
          format: date-time
        paymentMethod:
          type: string
          enum: [click, payme, paynet, cash]
      required:
        - toolId
        - customerId
        - dueDate
        - paymentMethod
    ReturnRentalRequest:
      type: object
      properties:
        damageNote:
          type: string
        damageCost:
          type: number
    StartRentalByQrRequest:
      type: object
      properties:
        qrCode:
          type: string
        customerId:
          type: number
        dueDate:
          type: string
          format: date-time
        paymentMethod:
          type: string
          enum: [click, payme, paynet, cash]
      required:
        - qrCode
        - customerId
        - dueDate
        - paymentMethod
    ReturnByQrRequest:
      type: object
      properties:
        qrCode:
          type: string
        damageNote:
          type: string
        damageCost:
          type: number
      required:
        - qrCode
    RentalListResponse:
      type: object
      properties:
        rentals:
          type: array
          items:
            $ref: "#/components/schemas/Rental"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - rentals
        - total
        - page
        - limit
...